backend/pyproject.toml
[project]
name = "rfp-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.4.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.12.0",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]
